Did the same changes from the SDL3 + GL impl to the SDL3 impl

+++ src/impl/sdl3.c
@@ -164,7 +164,13 @@ void impl_init(
 		handleEvent();
 		draw();
 		
-		SDL_Delay((1000.0f / MAX_FPS) - (SDL_GetTicks() - lastTime));
+		SDL_Delay(
+			(
+				((float)1000.0f) / MAX_FPS
+			) - (
+				SDL_GetTicks() - lastTime
+			)
+		);
 		
 		handleEvent();
 	}
